+++ HISTORY.md
@@ -3,6 +3,9 @@ History
 
 ## Unreleased
 
+* Add special `_templatesDir` prompt for specifying alternate templates
+  (`init/`) directory.
+  [#36](https://github.com/FormidableLabs/builder-init/issues/36)
 * Update lint, build, publish infrastructure.
 
 ## 0.2.4